Feature: Add project create tool (#3)
* Fix: Syntax issue in concurrency config for GHA workflow

Wrong identation was considering all concurrency related config environment
variables.

* Fix: Change URLs for staging environment

The staging environment has specific URLs for the OAuth flow.

* Feature: Add project create tool

This is the first design proof-of-concept of tools in the repository.

Task https://digitalcrew.teamwork.com/app/tasks/24833874

* Fix: Inject secrets now that the service depends on private repositories

This service will not rely on vendoring folder for now as the private
dependency may become public in a near future.